Move set_jwt_validation_policy into application code (#358)
