+++ test/system-test/test_all_seq.py
@@ -391,20 +391,10 @@ def test_set_policy_single_key_no_jwt_auth_jwt(setup_kms_session):
             raise
 
 
-@pytest.mark.xfail(
-    os.getenv("TEST_ENVIRONMENT") == "ccf/acl",
-    strict=True,
-    reason="Governance operations need to move to user endpoints",
-)
 def test_set_policy_single_key_no_jwt_trust_jwt_issuer(setup_kms_session, setup_aad_jwt_issuer_session):
     trust_jwt_issuer("aad")
 
 
-@pytest.mark.xfail(
-    os.getenv("TEST_ENVIRONMENT") == "ccf/acl",
-    strict=True,
-    reason="Governance operations need to move to user endpoints",
-)
 def test_set_policy_single_key_set_jwt_auth_jwt(setup_kms_session):
     status_code, auth_json = auth(auth="jwt")
     assert status_code == 200
