+++ test/system-test/test_auth.py
@@ -32,18 +32,13 @@ def test_auth_demo_jwt(setup_kms, setup_demo_jwt_issuer):
     assert auth_json["auth"]["policy"] == "jwt"
     assert auth_json["auth"]["jwt"]["keyIssuer"] == "http://Demo-jwt-issuer"
 
-@pytest.mark.xfail(
-    os.getenv("TEST_ENVIRONMENT") == "ccf/acl",
-    strict=True,
-    reason="Governance operations need to move to user endpoints",
-)
 def test_auth_aad_jwt(setup_kms, setup_aad_jwt_issuer):
     apply_kms_constitution()
     trust_jwt_issuer("aad")
     status_code, auth_json = auth(auth="jwt")
     assert status_code == 200
     assert auth_json["auth"]["policy"] == "jwt"
-    assert "https://sts.windows.net" in auth_json["auth"]["jwt"]["keyIssuer"]
+    assert "https://login.microsoftonline.com" in auth_json["auth"]["jwt"]["keyIssuer"]
